finish reorganisation
