+++ architrice/caching.py
@@ -5,149 +5,18 @@ import os
 import typing
 
 from . import database
+from . import deckreprs
 from . import sources
 from . import targets
 from . import utils
 
 
-class Card(database.StoredObject):
-    def __init__(
-        self,
-        name,
-        mtgo_id=None,
-        is_dfc=False,
-        collector_number=None,
-        edition=None,
-        db_id=None,
-    ):
-        super().__init__("cards", db_id)
-        self.name: str = name
-        self.mtgo_id: str = mtgo_id
-        self.is_dfc: bool = is_dfc
-        self.collector_number: str = collector_number
-        self.edition: str = edition
-
-    def __repr__(self):
-        return (
-            f"<Card name={self.name} mtgo_id={self.mtgo_id} "
-            f"is_dfc={self.is_dfc} collector_number={self.collector_number} "
-            f"edition={self.edition} id={self._id}>"
-        )
-
-    @staticmethod
-    def from_record(tup):
-        # database record format:
-        # (id, name, mtgo_id, is_dfc, collector_number, set, is_reprint)
-        _, name, mtgo_id, is_dfc, collector_number, edition, _ = tup
-        return Card(
-            name, mtgo_id and str(mtgo_id), is_dfc, collector_number, edition
-        )
-
-
-class DeckDetails(database.StoredObject):
-    """A DeckDetails object represents a deck in a source."""
-
-    def __init__(self, deck_id, source, db_id=None):
-        super().__init__("decks", db_id)
-        self.deck_id: str = deck_id
-        self.source: str = source
-
-    def __hash__(self):
-        return hash((self.deck_id, self.source))
-
-    def __repr__(self):
-        return (
-            f"<DeckDetails deck_id={self.deck_id} source={self.source} "
-            f"id={self._id}>"
-        )
-
-
-class Deck(DeckDetails):
-    """A Deck object represents a deck downloaded from a source."""
-
-    # The cards held by the deck are (quantity, name) tuples rather than Card
-    # objects. They are parsed into Cards before saving.
-    def __init__(self, deck_id, source, name, description, **kwargs):
-        super().__init__(deck_id, source, db_id=kwargs.get("id"))
-        self.name: str = name
-        self.description: str = description
-        self.main: typing.List[typing.Tuple[int, str]] = kwargs.get("main", [])
-        self.side: typing.List[typing.Tuple[int, str]] = kwargs.get("side", [])
-        self.maybe: typing.List[typing.Tuple[int, str]] = kwargs.get(
-            "maybe", []
-        )
-        self.commanders: typing.List[typing.Tuple[int, str]] = kwargs.get(
-            "commanders", []
-        )
-
-    def __repr__(self):
-        return super().__repr__().replace("<DeckDetails", "<Deck")
-
-    def get_card_names(self, board):
-        return [c[1] for c in self.get_board(board)]
-
-    def get_all_card_names(self):
-        return set(
-            self.get_card_names("main")
-            + self.get_card_names("side")
-            + self.get_card_names("maybe")
-            + self.get_card_names("commanders")
-        )
-
-    def get_main_deck(self, include_commanders=False):
-        if include_commanders:
-            return self.main + self.commanders
-        return self.main
-
-    def get_sideboard(self, include_commanders=True, include_maybe=True):
-        sideboard = self.side[:]
-        if include_commanders:
-            sideboard += self.commanders
-        if include_maybe:
-            sideboard += self.maybe
-        return sideboard
-
-    def get_board(self, board, default="main"):
-        board = board.strip().lower()
-        if board == "commanders":
-            return self.commanders
-        elif board in ["maybe", "maybeboard"]:
-            return self.maybe
-        elif board in ["side", "sideboard"]:
-            return self.side
-        elif board in ["main", "maindeck", "mainboard"]:
-            return self.main
-        else:
-            return self.get_board(default)
-
-    def add_card(self, card, board):
-        self.get_board(board).append(card)
-
-    def add_cards(self, cards, board):
-        self.get_board(board).extend(cards)
-
-
-class DeckUpdate:
-    """A DeckUpdate represents the last time a Deck was updated on a source."""
-
-    # Because these are not stored anywhere, they don't need a db id.
-    def __init__(self, deck, updated):
-        self.deck: DeckDetails = deck
-        self.updated: int = updated
-
-    def __repr__(self):
-        return f"<DeckUpdate deck={repr(self.deck)} updated={self.updated}>"
-
-    def update(self):
-        self.updated = utils.time_now()
-
-
-class DeckFile(database.StoredObject, DeckUpdate):
+class DeckFile(database.StoredObject, deckreprs.DeckUpdate):
     """A DeckFile represents the last time a local Deck file was updated."""
 
     def __init__(self, deck, updated, file_name, output, db_id=None):
         database.StoredObject.__init__(self, "deck_files", db_id)
-        DeckUpdate.__init__(self, deck, updated)
+        deckreprs.DeckUpdate.__init__(self, deck, updated)
         self.file_name: str = file_name
         self.output: Output = output
 
@@ -574,7 +443,7 @@ class Profile(database.StoredObject):
 
 
 class Cache:
-    def __init__(self, profiles=None, output_dirs=None):
+    def __init__(self, profiles=None):
         self.profiles: typing.List[Profile] = profiles or []
 
     def add_profile(self, profile):
@@ -599,6 +468,9 @@ class Cache:
 
     def build_output(self, profile, target, path, include_maybe):
         profile.add_output(Output(target, OutputDir.get(path), include_maybe))
+
+    def get_all_output_dirs(self):
+        return OutputDir.get_all()
 
     def save(self):
         logging.debug("Saving cache to database.")
@@ -645,6 +517,9 @@ class Cache:
         if source:
             conditions.append("u.source = ?")
             arguments.append(source.short)
+        if name:
+            conditions.append("p.name = ?")
+            arguments.append(name)
         if conditions:
             query += " WHERE " + " AND ".join(conditions)
         for tup in database.execute(query + ";", arguments):
@@ -699,7 +574,7 @@ class Cache:
                     output.output_dir.add_deck_file(
                         output,
                         DeckFile(
-                            DeckDetails(d_deck_id, d_source, d_db_id),
+                            deckreprs.DeckDetails(d_deck_id, d_source, d_db_id),
                             df_updated,
                             df_file_name,
                             output,
@@ -716,4 +591,4 @@ class Cache:
                 )
             )
 
-        return Cache(profiles, output_dirs)
+        return Cache(profiles)
