+++ architrice/modes/__init__.py
@@ -1,15 +1,19 @@
-from .add import Add
-from .delete import Delete
+from .profile import AddProfile
+from .delete import DeleteProfile
 from .edit import Edit
 from .latest import Latest
+from .output import AddOutput
 from .relnk import Relnk
 from .version import Version
 
-modelist = [
-    Add(),
-    Delete(),
+from .sync import Sync
+
+flag_modes = [
+    AddProfile(),
+    DeleteProfile(),
     Edit(),
     Latest(),
+    AddOutput(),
     Relnk(),
-    Version()
+    Version(),
 ]
