+++ architrice/modes/sync.py
@@ -3,16 +3,16 @@ import logging
 from . import mode
 
 
-class Latest(mode.FilterArgsMode):
+class Sync(mode.FilterArgsMode):
     def __init__(self):
-        super().__init__("l", "latest", "download latest deck for user")
+        super().__init__("s", "sync", "sync decklists", ["profiles"])
 
     def action(self, cache, args):
         if args.profiles:
             for profile in args.profiles:
-                profile.download_latest()
+                profile.download_all()
         else:
             logging.info(
                 "No matching profiles to sync."
-                " Add one with \"python -m architrice -a\"."
+                " Add one with \"python -m architrice -a\""
             )
