+++ architrice/modes/version.py
@@ -7,5 +7,5 @@ class Version(mode.Mode):
     def __init__(self):
         super().__init__("v", "version", "print version and exit")
 
-    def main():
+    def main(self, args):
         print(architrice.__version__)
