+++ architrice/sources/source.py
@@ -1,9 +1,8 @@
 import abc
-import dataclasses
 import logging
 
-from .. import caching
 from .. import database
+from .. import deckreprs
 
 
 class Source(database.KeyStoredObject, abc.ABC):
@@ -24,7 +23,7 @@ class Source(database.KeyStoredObject, abc.ABC):
 
     def create_deck(self, deck_id, name, description):
         """Create a Deck with relevant information."""
-        return caching.Deck(deck_id, self.short, name, description)
+        return deckreprs.Deck(deck_id, self.short, name, description)
 
     @abc.abstractmethod
     def _get_deck(self, deck_id):
@@ -81,3 +80,8 @@ class Source(database.KeyStoredObject, abc.ABC):
         else:
             logging.error("Verfification failed.")
         return result
+
+    def deck_update_from(self, deck_id, time):
+        return deckreprs.DeckUpdate(
+            deckreprs.DeckDetails(deck_id, self.short), time
+        )
